Multimodal_RAG_HANA/requirements.txt
absl-py==2.1.0
aenum==3.1.15
ai-api-client-sdk==2.1.0
ai-core-sdk==2.3.2
aiodns==3.2.0
aiofiles==23.2.1
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aleph-alpha-client==7.1.0
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
atlassian-python-api==3.41.14
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bert-score==0.3.13
boto3==1.34.117
botocore==1.34.117
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
ci-info==0.3.0
click==8.1.7
cohere==5.5.4
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.6
datasets==2.20.0
debugpy==1.8.1
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
effdet==0.4.1
email_validator==2.1.1
emoji==2.12.1
et-xmlfile==1.1.0
etelemetry==0.3.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.4
ffmpy==0.3.2
filelock==3.14.0
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==24.3.25
fonttools==4.53.0
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
generative-ai-hub-sdk==1.5.1
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-vision==3.7.3
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
gradio==4.36.1
gradio_client==1.0.1
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
hdbcli==2.21.21
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iopath==0.1.10
ipykernel==6.29.4
ipython==8.25.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==1.0.3
langchain-openai==0.1.6
langchain-text-splitters==0.0.2
langdetect==1.0.9
langfuse==2.39.3
langsmith==0.1.67
layoutparser==0.3.4
looseversion==1.3.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.0
open-clip-torch==2.24.0
openai==1.30.5
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.3
overloading==0.5.0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.2
pexpect==4.9.0
pikepdf==9.0.0
pillow==10.3.0
pillow_heif==0.17.0
platformdirs==4.2.2
portalocker==2.10.0
posthog==3.5.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
prov==2.0.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycares==4.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
pydot==2.0.0
pydub==0.25.1
Pygments==2.18.0
pyhumps==3.8.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pysbd==0.3.4
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-liquid==1.12.1
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==0.6.23
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==26.0.3
ragas==0.1.11
rapidfuzz==3.9.4
rdflib==6.3.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rouge_score==0.1.2
rpds-py==0.18.1
rsa==4.9
ruff==0.4.8
s3transfer==0.10.1
sacrebleu==2.4.2
safetensors==0.4.3
sap-llm-commons==0.2.0
scipy==1.14.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
tenacity==8.3.0
text-generation==0.7.0
tiktoken==0.5.2
timm==1.0.7
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
traits==6.3.2
transformers==4.42.3
typer==0.12.3
types-requests==2.32.0.20240602
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
ujson==5.10.0
unstructured==0.14.10
unstructured-client==0.24.0
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.12
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
